smartplanner/requirements.txt
Django==5.0.2
gunicorn==21.2.0
pandas==2.2.1
joblib==1.3.2
python-decouple==3.8
google-generativeai==0.3.2
